Two interesting updates: ...
In Quadcopter, the inverse square root function is replaced with a simpler and better one that quite fast enough on a Teensy 4.0 as well as more precise.

In Transmitter code, the ButtonWasPressed function now includes the first 12 functions invoked from an array of function pointers for greater speed. Currently text commands are still supported too, but eventually they must all be replaced.